requirements.txt
aiogram==2.22.2
aiohttp==3.8.3
aiosignal==1.2.0
art==5.7
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
branca==0.5.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
exceptiongroup==1.0.0
Fiona==1.8.22
folium==0.13.0
fonttools==4.38.0
frozenlist==1.3.1
geographiclib==1.52
geopandas==0.12.1
geopy==2.2.0
h11==0.14.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
multidict==6.0.2
munch==2.5.0
networkx==2.8.8
numpy==1.23.4
osmnx==1.2.2
outcome==1.2.0
packaging==21.3
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
pyparsing==3.0.9
pyproj==3.4.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
requests==2.28.1
Rtree==1.0.1
scikit-learn==1.1.3
scipy==1.9.3
selenium==4.5.0
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
threadpoolctl==3.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.4
wsproto==1.2.0
yarl==1.8.1